Removed export of local consumption constants in the agent config.
The environment folder has been moved to the internal.

Removed the decode in the configurator. The configurator section and file have been deprecated as empty.

Fixed the implementation of the cycle for collecting and sending metrics in DoTheJob to the select case, which explicitly takes into account the volatile nature of ReportInterval.

Added error return from GetHTMLPageInfo